+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix bug in L048 which flagged adjoining commas as failures.
 - Fix bug in L019 with inline comments.
 - Fix bug in L036 with multiple newlines.
+- Skip disabled dbt models. ([#931](https://github.com/sqlfluff/sqlfluff/issues/931)).
 - Support "USE" statement in ANSI ([#902](https://github.com/sqlfluff/sqlfluff/issues/902)).
 - Parse explain statement ([#893](https://github.com/sqlfluff/sqlfluff/issues/893)).